more component improvements
