+++ web/package.json
@@ -20,6 +20,7 @@
   "homepage": "https://github.com/rosnovsky/auth0-example-apps#readme",
   "private": true,
   "dependencies": {
+    "@reach/router": "^1.3.3",
     "auth0-js": "^9.13.2",
     "bootstrap": "^4.4.1",
     "core-js": "^3.6.5",
